MAJOR refactoring
Two major changes:
1) now accepts three parameters: plone_id, catalog_id and index_id.

This way you can select a specific plone instance to optimize,
or a catalog within a plone instance,
or an index within a catalog within a plone instance.

Parameters are cumulative: i.e one can not specify only a catalog or an index.

2) better progress reporting

More output is shown on the commandline and big objects (like indexes) are
splitted into batches for processing their trees, this way one can see how the
optimization is progressing.